axel: Fix speed limiting, and use 64-bit sizes and offsets
Reset delay_time only when it's too small for another subtraction,
the original code wrongly cleared it when the speed is within spec.

Variables used to overflow on 32-bit systems, use off_t instead, which
should be 64 bits on systems supporting that.

[ismael@iodev.co.uk: Improved commit message]

+++ src/axel.h
@@ -129,8 +129,8 @@ typedef struct {
 	char filename[MAX_STRING];
 	double start_time;
 	int next_state, finish_time;
-	size_t bytes_done, start_byte, size;
-	int bytes_per_second;
+	off_t bytes_done, start_byte, size;
+	long long int bytes_per_second;
 	struct timespec delay_time;
 	int outfd;
 	int ready;
